make an upgrade module where we'll be able to shove each version instead of putting everything in the same file
